+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Reduce the size of the struct returned by `UnicodeCategorySet.iter()` from 8 to 4 bytes and improve performance for
   cases when a few Unicode categories are involved.
+- Reduce allocations inside `query` for cases when a few Unicode categories are involved.
 
 ## [0.1.2] - 2023-04-23